+++ src/main/java/com/qa/persistence/domain/Trainer.java
@@ -6,9 +6,8 @@ public class Trainer extends User {
 		this.setFirstName(trainee.getFirstName());
 		this.setLastName(trainee.getLastName());
 		this.setUsername(trainee.getUsername());
-		this.setPassword(trainee.getPassword());
 	}
-	
+
 	public Trainer(String username) {
 		super(username);
 		this.type = "Trainer";
